fix: harden catalog herdr bridge and verification
