+++ catalog/packages/herdr-bridge/colorize.py
@@ -1,13 +1,7 @@
 #!/usr/bin/env python3
-"""Colore e enxuga o stream de logs do CompozyOS para o pane do herdr.
+"""Render original Compozy session events and JSONL logs for herdr panes.
 
-Renderiza eventos originais via tail.py e aceita logs JSONL no stdin.
-Cor e filtragem nascem aqui; mensagens usam content.text sem limpeza.
-
-Medido num loop real (400 eventos): 30% era `usage`, 15% era `tool_result`
-com summary "[REDACTED]" (100% deles), 14% era linha repetida identica e
-4% era chatter de resolucao de skill. Menos de um terco tinha conteudo.
-"""
+Preserve message whitespace while filtering infrastructure noise."""
 import json
 import re
 import shutil
@@ -15,7 +9,7 @@ import sys
 
 RESET = "\033[0m"
 
-# ruido de infraestrutura: nao diz nada sobre o que o agente esta fazendo
+# Infrastructure noise does not explain agent activity.
 SKIP_TYPES = {
     "usage",
     "session_stream_subscribed",
@@ -25,11 +19,11 @@ SKIP_TYPES = {
 }
 SKIP_TYPE_PREFIXES = ("harness.",)
 
-# resultado de ferramenta vem sempre censurado; a linha so ocupa espaco
+# Empty or redacted results add no useful output.
 EMPTY_SUMMARIES = {"[REDACTED]", "", "null"}
 
-# rotulo de ferramenta ("Terminal", "Edit"...) vem como evento proprio,
-# seguido do comando de verdade. Junta os dois numa linha so.
+# Tool labels arrive before the command itself.
+# Combine both into one rendered line.
 TOOL_LABEL = re.compile(r"^[A-Z][A-Za-z]{2,19}$")
 
 TYPE_STYLE = {
@@ -56,7 +50,14 @@ OUTCOME_STYLE = {
 }
 
 
+def terminal_safe(text):
+    """Escape terminal controls while preserving printable text, tabs, and newlines."""
+    return "".join(char if char.isprintable() or char in "\n\t"
+                   else f"\\x{ord(char):02x}" for char in text)
+
+
 def style_for(etype, outcome):
+    """Select a trusted ANSI style for the event and outcome."""
     if outcome in OUTCOME_STYLE:
         return OUTCOME_STYLE[outcome]
     if etype in TYPE_STYLE:
@@ -68,18 +69,19 @@ def style_for(etype, outcome):
 
 
 def width():
+    """Return the usable terminal width with a conservative fallback."""
     try:
         return max(60, shutil.get_terminal_size().columns) - 2
     except Exception:
         return 118
 
 
-# todo comando do agente vem como `cd /caminho/longo && <o que importa>`
+# Remove the working-directory prefix from command summaries.
 CD_PREFIX = re.compile(r"^cd\s+\S+\s*(?:&&|;)\s*")
 
 
 def flatten(summary):
-    """Heredoc de varias linhas vira uma linha; o corpo raramente importa."""
+    """Summarize a multiline command using its first nonempty line."""
     parts = [x.strip() for x in summary.replace("\r", "").split("\n") if x.strip()]
     if not parts:
         return ""
@@ -88,7 +90,7 @@ def flatten(summary):
 
 
 def session_event(event):
-    """Adapta o evento original sem confundir seu texto com resumos de logs."""
+    """Adapt original events without treating message text as log summaries."""
     content = event.get("content")
     if not isinstance(content, dict) or "sequence" not in event:
         return event
@@ -98,8 +100,8 @@ def session_event(event):
     if etype == "agent_message":
         summary = content.get("text", "")
     elif etype == "tool_result":
-        # A API original inclui resultados completos. Mantem o filtro de ruido
-        # sem despejar tool_input/tool_result no terminal.
+        # Original events include full tool results; preserve noise filtering.
+        # Do not dump tool inputs or successful result bodies.
         summary = (content.get("error") or content.get("text") or
                    content.get("title") or "tool failed") if error else ""
     else:
@@ -110,19 +112,22 @@ def session_event(event):
 
 
 class Renderer:
+    """Render event fragments with per-session streaming and deduplication."""
     def __init__(self):
-        self.last_text = ""         # ultima linha impressa, pra reescrever igual
-        self.pending_label = None   # rotulo de ferramenta esperando o comando
-        self.streaming = False      # dentro de uma sequencia de agent_message
-        self.stream_scope = None    # sessao/turno; mensagens irmas nao se fundem
-        self.last_key = None        # (tipo, texto) da ultima linha impressa
+        """Initialize streaming, tool label, and repeated-line state."""
+        self.last_text = ""         # Last rendered line for repeat updates.
+        self.pending_label = None   # Tool label awaiting a command.
+        self.streaming = False      # Inside an agent_message stream.
+        self.stream_scope = None    # Keep sibling sessions and turns separate.
+        self.last_key = None        # Last rendered event type and text.
         self.repeat = 1
 
     def out(self, text, key=None):
+        """Emit trusted styled text and collapse repeated rendered lines."""
         if key is not None and key == self.last_key:
-            # repetida: reescreve a propria linha com o contador
-            # a linha anterior ja terminou com \n, entao sobe uma linha,
-            # limpa e reescreve com o contador
+            # Rewrite the previous line with its repeat count.
+            # Move up after the previous newline.
+            # Clear and redraw the line with a counter.
             self.repeat += 1
             sys.stdout.write(f"\033[A\r\033[K{self.last_text} \033[2;90m×{self.repeat}{RESET}\n")
             sys.stdout.flush()
@@ -134,6 +139,7 @@ class Renderer:
         sys.stdout.flush()
 
     def close_stream(self):
+        """Finish the active message stream before a different event."""
         if self.streaming:
             sys.stdout.write("\n")
             sys.stdout.flush()
@@ -141,27 +147,28 @@ class Renderer:
             self.last_key = None
 
     def feed(self, d):
+        """Render one event while preserving message fragment boundaries."""
         d = session_event(d)
-        etype = str(d.get("type") or "?")
+        etype = terminal_safe(str(d.get("type") or "?"))
         if etype in SKIP_TYPES or etype.startswith(SKIP_TYPE_PREFIXES):
             return
         outcome = str(d.get("outcome") or "")
-        summary = str(d.get("summary") or "")
-        # Deltas de mensagem carregam os espacos entre palavras. A limpeza
-        # de comandos remove essas bordas e tambem destroi quebras/indentacao.
+        summary = terminal_safe(str(d.get("summary") or ""))
+        # Message deltas include spaces between words.
+        # Command flattening would destroy message breaks and indentation.
         if etype != "agent_message":
             summary = flatten(summary)
 
-        # resultado vazio nao vira linha (a menos que seja falha)
+        # Suppress empty successful tool results.
         if etype == "tool_result" and summary in EMPTY_SUMMARIES and outcome not in OUTCOME_STYLE:
             return
 
-        # rotulo de ferramenta: segura pro proximo evento
+        # Hold the tool label for the following command.
         if etype == "tool_call" and TOOL_LABEL.match(summary):
             self.pending_label = summary
             return
 
-        ts = str(d.get("timestamp") or "")[11:19]
+        ts = terminal_safe(str(d.get("timestamp") or "")[11:19])
         code = style_for(etype, outcome)
 
         if etype == "agent_message":
@@ -194,6 +201,7 @@ class Renderer:
 
 
 def main():
+    """Render JSONL events and plain fallback lines from standard input."""
     r = Renderer()
     for line in sys.stdin:
         line = line.strip()
@@ -201,7 +209,7 @@ def main():
             continue
         if not line.startswith("{"):
             r.close_stream()
-            r.out(f"\033[2;37m{line}{RESET}")
+            r.out(f"\033[2;37m{terminal_safe(line)}{RESET}")
             continue
         try:
             r.feed(json.loads(line))
