+++ catalog/packages/herdr-bridge/tail.py
@@ -1,9 +1,7 @@
 #!/usr/bin/env python3
-"""Segue os eventos originais das sessoes de uma linha do bridge.
+"""Follow original session events for one bridge row.
 
-Logs sao resumos aparados e truncados, nao deltas de texto. O cursor de
-sequencia da API de sessoes permite acompanhar o conteudo sem perder espacos.
-"""
+Sequence cursors preserve message whitespace and resume after reconnects."""
 import argparse
 import http.client
 import json
@@ -11,7 +9,7 @@ import socket
 import time
 from urllib.parse import quote, urlencode
 
-import bridge
+import bridge_state as bridge
 from colorize import Renderer
 
 POLL_SECONDS = 1
@@ -19,6 +17,7 @@ INITIAL_EVENTS = 100
 
 
 def get_json(path):
+    """Read one JSON response from the local daemon socket."""
     connection = http.client.HTTPConnection("localhost", timeout=5)
     connection.sock = socket.socket(socket.AF_UNIX, socket.SOCK_STREAM)
     connection.sock.settimeout(5)
@@ -34,7 +33,9 @@ def get_json(path):
 
 
 class SessionTail:
+    """Track event cursors for the live sessions belonging to one row."""
     def __init__(self, key, renderer):
+        """Initialize row scope and per-session cursors."""
         self.key = key
         self.workspace = key.split("/", 1)[0]
         self.renderer = renderer
@@ -42,6 +43,7 @@ class SessionTail:
         self.errors = {}
 
     def poll(self, data):
+        """Read new events and drain removed sessions before dropping cursors."""
         sessions = data.get(self.key, {}).get("sessions", {})
         for sid in sessions:
             self.cursors.setdefault(sid, None)
@@ -50,15 +52,21 @@ class SessionTail:
             if cursor is None:
                 query["limit"] = INITIAL_EVENTS
             else:
-                # limit significa os ULTIMOS N, mesmo com after_sequence.
-                # Limitar aqui descartaria o inicio de rajadas ou reconexoes.
+                # limit selects the LAST N even with after_sequence.
+                # Do not truncate bursts or reconnect backlogs.
                 query["after_sequence"] = cursor
-            path = (f"/api/workspaces/{quote(self.workspace, safe='')}/sessions/"
-                    f"{quote(sid, safe='')}/events?{urlencode(query)}")
             try:
+                workspace = bridge.normalize_workspace(self.workspace)
+                if not workspace:
+                    owner = get_json(f"/api/sessions/{quote(sid, safe='')}/owner")
+                    workspace = owner["workspace_id"]
+                    if not workspace or owner["session_id"] != sid:
+                        raise ValueError("session owner response does not match the requested session")
+                path = (f"/api/workspaces/{quote(workspace, safe='')}/sessions/"
+                        f"{quote(sid, safe='')}/events?{urlencode(query)}")
                 events = get_json(path)["events"]
             except (OSError, http.client.HTTPException, ValueError, KeyError) as exc:
-                # Uma falha nao avanca o cursor nem interrompe sessoes irmas.
+                # Failures neither advance the cursor nor interrupt sibling sessions.
                 if self.errors.get(sid) != str(exc):
                     bridge.log(f"tail {sid}: {exc}")
                 self.errors[sid] = str(exc)
@@ -71,14 +79,15 @@ class SessionTail:
                     self.cursors[sid] = sequence
             if self.cursors[sid] is None:
                 self.cursors[sid] = 0
-            # Le a ultima leva de uma sessao removida antes de esquecer o cursor.
+            # Read the final batch before forgetting a removed session.
             if sid not in sessions:
                 self.cursors.pop(sid)
 
 
 def main():
+    """Poll the selected row and flush its renderer on exit."""
     parser = argparse.ArgumentParser(description=__doc__)
-    parser.add_argument("row", help="Chave workspace_id/agent_name em panes.json")
+    parser.add_argument("row", help="workspace_id/agent_name key in panes.json")
     args = parser.parse_args()
     renderer = Renderer()
     tail = SessionTail(args.row, renderer)
