+++ docs/qa/charters/CH-herdr-bridge-runtime.md
@@ -1,0 +1,11 @@
+# CH-herdr-bridge-runtime
+
+- Mission: As Ada, install the bridge and follow a session lifecycle in herdr without losing rows or exposing terminal controls.
+- Persona: Ada, technical extension author using macOS CLI, local network, English locale.
+- Journey: J-extension-distribution.
+- Tour: Feature.
+- Time-box: 30 minutes.
+- Scenarios: ET-herdr-bridge-runtime.
+- Entry: catalog CLI production installation, extension hook stdin, bridge status, herdr public pane API.
+- End state: completed session pane closes, unrelated panes survive, isolated lab tears down cleanly.
+- Adjacent canary: standalone event renderer preserves normal message whitespace.
